requirements.txt
Django>=1.9.6
django-allauth
